+++ include/sonia_bt_runner/SoniaNodes.hpp
@@ -15,7 +15,7 @@ static void registerSyncActionNodes(BT::BehaviorTreeFactory &factory)
     factory.registerNodeType<DropperActionNode>("DropperActionNode");
     factory.registerNodeType<InitTraj>("InitTraj");
     factory.registerNodeType<AddPoseToTraj>("AddPoseToTraj");
-    factory.registerNodeType<AIActivation>("AIActivation");
+    factory.registerNodeType<AiActivation>("AiActivation");
     factory.registerNodeType<GetAiInfo>("GetAiInfo");
     factory.registerNodeType<EnableDisableDVL>("EnableDisableDVL");
     factory.registerNodeType<TareDepth>("TareDepth");
