+++ include/sonia_bt_runner/stateful_actions/SendTrajToPlanner.hpp
@@ -11,13 +11,8 @@ using namespace std::chrono_literals;
 class SendTrajToPlanner : public BT::StatefulActionNode
 {
 public:
-    SendTrajToPlanner(const std::string &name, const BT::NodeConfig &config)
-        : BT::StatefulActionNode(name, config), _valid(0), _time_launch(std::chrono::system_clock::now())
-    {
-
-        _planner_pub = _nh.advertise<sonia_common::MultiAddPose>("/proc_planner/send_multi_addpose", 1, true);
-    }
-
+    SendTrajToPlanner(const std::string &name, const BT::NodeConfig &config);
+    
     static BT::PortsList providedPorts()
     {
         const int def_inter = 0;
