+++ include/sonia_bt_runner/sync_actions/AddPoseToTraj.hpp
@@ -5,10 +5,7 @@
 class AddPoseToTraj : public BT::SyncActionNode
 {
 public:
-    AddPoseToTraj(const std::string &name, const BT::NodeConfig &config)
-        : BT::SyncActionNode(name, config)
-    {
-    }
+    AddPoseToTraj(const std::string &name, const BT::NodeConfig &config);
 
     static BT::PortsList providedPorts()
     {
