+++ include/sonia_bt_runner/sync_actions/SoniaSyncActions.hpp
@@ -3,7 +3,7 @@
 #include "sonia_bt_runner/sync_actions/DropperActionNode.hpp"
 #include "sonia_bt_runner/sync_actions/InitTraj.hpp"
 #include "sonia_bt_runner/sync_actions/AddPoseToTraj.hpp"
-#include "sonia_bt_runner/sync_actions/AIActivation.hpp"
+#include "sonia_bt_runner/sync_actions/AiActivation.hpp"
 #include "sonia_bt_runner/sync_actions/GetAiInfo.hpp"
 #include "sonia_bt_runner/sync_actions/TareDepth.hpp"
 #include "sonia_bt_runner/sync_actions/EnableDisableDVL.hpp"
