+++ src/conditions/MissionSwitchStatus.cpp
@@ -1,5 +1,11 @@
 #include "sonia_bt_runner/conditions/MissionSwitchStatus.hpp"
 
+MissionSwitchStatus::MissionSwitchStatus(const std::string &name)
+    : BT::ConditionNode(name, {}), _mission_status(false)
+{
+    _mission_switch_sub = _nh.subscribe("/provider_kill_mission/mission_switch_msg", 10, &MissionSwitchStatus::update_status, this);
+}
+
 BT::NodeStatus MissionSwitchStatus::tick()
 {
     if (_mission_status)
