+++ src/stateful_actions/WaitTargetReached.cpp
@@ -1,5 +1,12 @@
 #include "sonia_bt_runner/stateful_actions/WaitTargetReached.hpp"
 
+WaitTargetReached::WaitTargetReached(const std::string &name, const BT::NodeConfig &config)
+    : BT::StatefulActionNode(name, config), _trajectory_done_prev(true),
+      _traj_complete(false), _launch_time(std::chrono::system_clock::now()), _param_timeout(0)
+{
+    _timeout_pub = _nh.advertise<sonia_common::MissionTimer>("/sonia_behaviors/timeout", 5);
+}
+
 BT::NodeStatus WaitTargetReached::onStart()
 {
     ROS_INFO("Starting WaitTargetReached");
