+++ src/sync_actions/EnableDisableDVL.cpp
@@ -1,5 +1,11 @@
 #include "sonia_bt_runner/sync_actions/EnableDisableDVL.hpp"
 
+EnableDisableDVL::EnableDisableDVL(const std::string &name, const BT::NodeConfig &config)
+    : BT::SyncActionNode(name, config)
+{
+    _dvl_pub = _nh.advertise<std_msgs::Bool>("/provider_dvl/enable_disable_ping", 100);
+}
+
 BT::NodeStatus EnableDisableDVL::tick()
 {
     std_msgs::Bool dvl_state;
